+++ frontend/src/App.jsx
@@ -4,33 +4,17 @@ import Footer from "./sections/Footer/Footer.jsx";
 import { useState, useEffect } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "./pages/Home.jsx";
-import Portfolio from "./pages/Portfolio.jsx";
-import PortfolioEntry from "./pages/PortfolioEntry.jsx";
+import Entry from "./pages/Entry.jsx";
 import NoPage from "./pages/NoPage.jsx";
 
 function App() {
-  const [portfolioEntries, setPortfolioEntries] = useState([]);
-
-  useEffect(() => {
-    async function fetchAllPortfolioEntries() {
-      await fetch("http://localhost:5000/portfolioEntries")
-        .then((res) => res.json())
-        .then((data) => {
-          setPortfolioEntries(data);
-        });
-    }
-
-    fetchAllPortfolioEntries();
-  }, []);
-
   return (
     <>
       <Navbar />
       <Router>
         <Routes>
           <Route index element={<Home />} />
-          <Route path="/portfolio" element={<Portfolio />} />
-          <Route path="/portfolioEntry/:ID" element={<PortfolioEntry />} />
+          <Route path="/Entry/:ID" element={<Entry />} />
           <Route path="*" element={<NoPage />} />
         </Routes>
       </Router>
