+++ frontend/src/pages/Home.jsx
@@ -1,7 +1,38 @@
+import { useState, useEffect } from "react";
+
 export default function Home() {
+  const [portfolioEntries, setPortfolioEntries] = useState([]);
+
+  useEffect(() => {
+    async function fetchAllPortfolioEntries() {
+      await fetch("http://localhost:5000/portfolioEntries")
+        .then((res) => res.json())
+        .then((data) => {
+          setPortfolioEntries(data);
+        });
+    }
+
+    fetchAllPortfolioEntries();
+  }, []);
+
   return (
     <>
-      <h1>hello, this is the home page</h1>
+      <main>
+        <h1>Entries</h1>
+        <ul>
+          {(!!portfolioEntries && portfolioEntries).map((n) => {
+            return (
+              <li key={n.ID}>
+                <a href={"/Entry/" + n.ID}>
+                  <h1>{n.title}</h1>
+                </a>
+                <p>{n.description}</p>
+                <p>{n.ID}</p>
+              </li>
+            );
+          })}
+        </ul>
+      </main>
     </>
   );
 }
